refactor: web側のAPIクライアントをrepositoryパターンに移行
lib/api/ から repository/ ディレクトリに移行し、server側と構成を統一

- thread-repository.ts: スレッド操作
- knowledge-repository.ts: ナレッジ管理
- persona-repository.ts: ペルソナ取得
- emergency-repository.ts: 緊急情報取得
- lib/api/client.ts: 共通APIクライアント（維持）
