+++ web/src/pages/chat/components/ChatContainer.tsx
@@ -4,8 +4,8 @@ import { DefaultChatTransport } from "ai";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { threadKeys, useCreateThread, useThreads } from "~/hooks/useThreads";
 import { API_BASE } from "~/lib/api/client";
-import { fetchMessages } from "~/lib/api/threads";
 import { getResourceId } from "~/lib/resource";
+import { fetchMessages } from "~/repository/thread-repository";
 import type { Thread } from "~/types";
 import { ChatInput } from "./ChatInput";
 import { MessageList } from "./MessageList";
